% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assemblage.R
\name{get_assemblage}
\alias{get_assemblage}
\title{Binary assemblage data generator}
\usage{
get_assemblage(obs, events, all_taxa = NULL)
}
\arguments{
\item{obs}{A vector of observations (e.g. taxonomic names)}

\item{events}{A vector of sampling events (i.e. sample IDs)}

\item{all_taxa}{Optional parameter to input a vector of all the unique taxa names of interest if they
are not all represented in \code{obs}. Default = NULL}
}
\value{
A data frame.
}
\description{
The \code{get_assemblage} function takes a vector of observations and their corresponding sampling events
and outputs a data frame with binary assemblage data for each sampling event.
}
