% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights}
\alias{get_weights}
\title{Calculates weights for classification mask}
\usage{
get_weights(x, y, all_taxa = NULL)
}
\arguments{
\item{x}{A data frame containing the ground truth assemblages.}

\item{y}{A data frame containing the detected assemblages.}

\item{all_taxa}{Optional parameter to input a vector of all the unique taxa names of interest if they
are not all represented in column names of \code{x}. Default = NULL.}
}
\value{
A data frame.
}
\description{
The \code{get_weights()} function takes two assemblage data frames (generated using the \code{get_assemblage()}
function) as input and outputs a data frame with the weighted classification mask weights for each
taxon in the assemblages. One input data frame should represent the ground truth assemblages and the
other input data frame should represent the detected assemblages (e.g. the assemblages detected using
DNA metabarcoding).
}
