% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longhier.R
\name{longhier}
\alias{longhier}
\title{Long hierarchy data frame generator.}
\usage{
longhier(x, hierarchy)
}
\arguments{
\item{x}{A character vector of taxonomic names. These names must be found in the first column of
the reference hierarchy data frame.}

\item{hierarchy}{A reference hierarchy data frame. This can be generated using the \code{hierarchy}
function.}
}
\value{
A data frame.
}
\description{
The \code{longhier} function takes a character vector of taxonomic names as input and cross references
them with a reference hierarchy data frame to create a long hierarchy data frame. This can be
useful for getting coarser grained taxonomic names from the class outputs of a classification
model.
}
