% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelbias.R
\name{modelbias}
\alias{modelbias}
\title{Model-biased classification refiner.}
\usage{
modelbias(dna_df, og_classes, og_levels, samples, agreed, hierarchy)
}
\arguments{
\item{dna_df}{Data frame of the DNA detections. The data frame must have these columns:
\itemize{
\item \code{sample_id}: The name of the sample the detection came from.
\item \code{known_class}: The class label known by the model (can be \code{NA} if it is not known by the model).
\item Taxonomic levels: Columns with the taxonomic name of the detection at each level. For example,
the \code{Species} column would have the species name of the detection (e.g., Gryllus rubens), and
the \code{Order} column would have the order level name of the detection (e.g., Orthoptera).
}}

\item{og_classes}{The original specimen classifications.}

\item{og_levels}{The original taxonomic ranks of the specimen classifications.}

\item{samples}{The sample IDs of the specimens.}

\item{agreed}{A logical vector that indicates if each classification was detected by the DNA in its
corresponding sample.}

\item{hierarchy}{A hierarchy data frame. This can be generated using the \code{hierarchy} function.}
}
\value{
A list with refined classifications and levels.
}
\description{
The \code{modelbias} function refines computer vision classifications by cross-referencing them
with DNA metabarcoding detections. It differs from the \code{dnabias} function in how it handles
disagreements between the classifications and the DNA metabarcoding detections.
}
