% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refhier.R
\name{refhier}
\alias{refhier}
\title{Reference hierarchy data frame generator.}
\usage{
refhier(df, ranks, base_rank, det_level = "Det_Level")
}
\arguments{
\item{df}{A data frame.}

\item{ranks}{A character vector of the taxonomic ranks to be included in the hierarchy. These will
be the column names of the output data frame.}

\item{base_rank}{The name of the column in \code{df} that contains the base taxa names (e.g. "Species")}

\item{det_level}{The name of the column in df that contains the detection levels. Detection levels
are the finest taxonomic rank that an observation was classified to. Any given value within the
column must correspond to the name of a column in \code{df}.}
}
\value{
A data frame.
}
\description{
The \code{refhier} function takes a data set with taxonomic information for individual observations
as input and outputs a reference hierarchy data frame. The hierarchy data frame contains taxonomic
information for all unique taxa in the input data frame.

If the taxonomic name for a given observation is missing at one or more ranks, the value for that
rank will be copied from the closest, coarser taxonomic rank with a name. For example, if an
observed species does not have an infraorder or suborder name, but it does have an order name,
the infraorder and suborder values for that species will match the order level name, assuming all
three ranks are included in \code{ranks}.
}
